test(lotto): separate lotto system tests by its' method steps
테스트코드 : LottoSystem의 메소드 진행 단계별로 구분하여 테스트 클래스를 분리
	- LottoSystem1 : purchaseLotto() 메소드까지 테스트 (로또 구매)
	- LottoSystem2 : drawLotto() 메소드까지 테스트 (로또 구매 - 추첨)
	- LottoSystem3 : getLottoResult() 메소드까지 테스트 (로또 구매 - 추첨 - 결과 산출)
