+++ src/test/java/lotto/controller/LottoSystemTest1.java
@@ -1,0 +1,51 @@
+package lotto.controller;
+
+import camp.nextstep.edu.missionutils.test.NsTest;
+import lotto.model.Lotto;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueNumbersInRangeTest;
+import static org.assertj.core.api.Assertions.assertThat;
+
+class LottoSystemTest1 extends NsTest {
+    private final LottoSystem lottoSystem = new LottoSystem();
+
+    @Test
+    @DisplayName("로또 구매하기")
+    void purchaseLottoTest1() {
+        assertRandomUniqueNumbersInRangeTest(
+                () -> {
+                    runException("8000");
+                    assertThat(output()).contains(
+                            "8개를 구매했습니다.",
+                            "[8, 21, 23, 41, 42, 43]",
+                            "[3, 5, 11, 16, 32, 38]",
+                            "[7, 11, 16, 35, 36, 44]",
+                            "[1, 8, 11, 31, 41, 42]",
+                            "[13, 14, 16, 38, 42, 45]",
+                            "[7, 11, 30, 40, 42, 43]",
+                            "[2, 13, 22, 32, 38, 45]",
+                            "[1, 3, 5, 14, 22, 45]"
+                    );
+                },
+                List.of(8, 21, 23, 41, 42, 43),
+                List.of(3, 5, 11, 16, 32, 38),
+                List.of(7, 11, 16, 35, 36, 44),
+                List.of(1, 8, 11, 31, 41, 42),
+                List.of(13, 14, 16, 38, 42, 45),
+                List.of(7, 11, 30, 40, 42, 43),
+                List.of(2, 13, 22, 32, 38, 45),
+                List.of(1, 3, 5, 14, 22, 45)
+        );
+    }
+
+    @Override
+    protected void runMain() {
+        List<Lotto> userLottos = lottoSystem.purchaseLotto();
+        List<?> wonLottoInfo = lottoSystem.drawLotto();
+        lottoSystem.getLottoResult(userLottos, wonLottoInfo);
+    }
+}
